+++ bot/SheepBot.Repositories/Interfaces/IRepositoryBase.cs
@@ -3,11 +3,11 @@ namespace SheepBot.Repositories.Interfaces;
 public interface IRepositoryBase<TModel>
 {
     Task<IEnumerable<TModel>> GetAllAsync();
-    Task<TModel?> FindAsync(int id);
+    Task<TModel?> FindAsync(long id);
     Task<IEnumerable<TModel>> FindAsync(Predicate<TModel> predicate);
-    Task<int> InsertAsync(TModel entity);
-    Task<int> InsertRangeAsync(IEnumerable<TModel> entities);
+    Task<long> InsertAsync(TModel entity);
+    Task<long> InsertRangeAsync(IEnumerable<TModel> entities);
     Task<int> UpdateAsync(TModel entity);
-    Task<int> DeleteAsync(int id);
+    Task<int> DeleteAsync(long id);
     Task<int> DeleteAsync(TModel entity);
 }
