+++ database/dbo/Tables/car.sql
@@ -1,5 +1,5 @@
 CREATE TABLE [dbo].[car] (
-    [id]        INT           NOT NULL IDENTITY(1,1),
+    [id]        BIGINT        NOT NULL IDENTITY(1,1),
     [name]      NVARCHAR (50) NOT NULL,
     [is_free]   BIT           NOT NULL,
     [is_legacy] BIT           NOT NULL
