+++ database/dbo/Tables/class_car.sql
@@ -1,7 +1,7 @@
 CREATE TABLE [dbo].[class_car] (
-    [id]       INT IDENTITY (1, 1) NOT NULL,
-    [class_id] INT NOT NULL,
-    [car_id]   INT NOT NULL,
+    [id]       BIGINT IDENTITY (1, 1) NOT NULL,
+    [class_id] BIGINT NOT NULL,
+    [car_id]   BIGINT NOT NULL,
     CONSTRAINT [PK_class_cars] PRIMARY KEY CLUSTERED ([id] ASC),
     CONSTRAINT [FK_class_cars_cars] FOREIGN KEY ([car_id]) REFERENCES [dbo].[car] ([id]),
     CONSTRAINT [FK_class_cars_class] FOREIGN KEY ([class_id]) REFERENCES [dbo].[class] ([id])
