+++ database/dbo/Tables/role.sql
@@ -1,5 +1,5 @@
 CREATE TABLE [dbo].[role] (
-    [id]         INT           NOT NULL IDENTITY(1,1),
+    [id]         BIGINT        NOT NULL IDENTITY(1,1),
     [discord_id] BIGINT        NOT NULL,
     [role_name]  NVARCHAR (20) NULL
 );
